ai-deps/requirements.txt
python-multipart==0.0.20
fastapi==0.115.12
fastapi-cli==0.0.7
pillow==11.2.1
numpy==2.1.3
transformers